Add UI design implementation and Runway configuration setup
Updated the HelloController and HelloWorldView for a more interactive UI and added more visual styles. The HelloWorldView now uses FXML file loaded by FXMLLoader instead of JavaFX components directly. A new Runway class was added, and Runway configurations are loaded from a text file to populate instances of this class. The UI contains more functionalities and the code has been given cleaner organization. A UI design XML was also included to manage the physical layout of the new interface components.