Add XML import feature to AirportList
This update introduces a new button for importing airports from an XML file in the AirportListScene's UI. When clicked, it opens a file chooser for the user to select the XML file. It then imports the airports, updates the app's airport list, and refreshes the displayed list. Changes were also made to the MainApplication class to facilitate the update of the app's airport list.