Update ObstacleListScene with new options and fixes
Several updates were made to the ObstacleListScene. New buttons for "Modify" and "View" were added to offer more options to the user. The "Remove" button was also corrected to correctly depict its function. There were also significant enhancements to the "Modify" functionality, which now includes a form for modifying the properties of an existing obstacle. Finally, the list of buttons in RunwayListScene and AirportListScene was rearranged for better user experience.