Update view scenes, refactor notifications and xml files
This commit adds the ability for the Runway Configuration View to close when returning to the Airport List Scene, and replaces the call to displayViewsScene with displayViewsSceneBeta in all instances. The Obstacle, Airport, and XML files have been updated to reflect these changes. Additionally, instant feedback has been provided to users when creating or reconfiguring obstacles, and when adding, merging, or updating airports.