Add background music to MainApplication
The MainApplication now includes a background music feature which is implemented by adding 'javafx.media' to the module-info. In detail, a new private method 'playBackgroundMusic', is created in MainApplication class. The method utilizes Media and MediaPlayer from javafx.scene.media package to play a sound file named 'background_music.mp3' that runs on loop as the application is running.