Add city image to SideViewScene, update logos
The commit introduces an aesthetic improvement to the SideViewScene by adding a city image as the background. The image is added with an opacity of 0.3 and positioned at the top center. The runway canvas and city image are stacked on top of each other. Furthermore, logos were added to the primary and secondary stages in the MainApplication.